requirements.txt
numpy==1.21.6
pandas==1.3.5
Pillow==9.0.0
img2dataset==1.28.0
img2vec_pytorch==1.0.1
torchvision
tqdm==4.63.0